Refactor and add rendering thread slider
