+++ include/CameraMessage.h
@@ -1,15 +1,18 @@
 #pragma once
-#include "WorldObject.h"
+
+#include "RtMathHelp.h"
 
 #include <string>
 
 struct CameraMessage
 {
+	int mNumRenderThreads;
 	double mIlluminationPercentage;
 	std::string mObjFilepath;
 	double mXAngle;
 	double mYAngle;
 	double mZAngle;
+
 
 	point3d mTranslation;
 };
